Make glfw notice when X11 DPI changes
glfw reads the DPI from the old X resources database attached to the
root window via the RESOURCE_MANAGER property. We just watch that
property, and when it changes, we re-read the database and potentially
react to any DPI changes. Bypassing XResourceManagerString and reading
the X property directly is a bit hackish, but the alternative is to
use xsettings, and using xsettings would require a new
external dependency.