Add support for high precision scrolling
On platforms such as macOS and Wayland scroll events report deltas in
pixels. glfw now supports this. This is a backward incompatible change.
The signature of the scroll callback function has changed and the values
of the offsets now must be interpreted differently. See the docs of
the callback function for details.